# openapi.yaml
openapi: 3.0.3
...
info:
  title: Adiaphora Bankruptcy Platform API
  description: >-
    Document-preparation platform for Russian personal-bankruptcy cases.
    Stateless JSON API under `/api/v1`; authentication is bearer JWT.
  version: 0.1.0
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/applications/{applicationId}/creditors:
    parameters:
      - $ref: '#/components/parameters/ApplicationIdPath'
    get:
      tags: [Creditors]
      summary: List creditors for a case
      responses:
        '200':
          description: Creditors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CreditorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Creditors]
      summary: Add a creditor (duplicates are flagged, not rejected)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreditorRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditorResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/applications/{applicationId}/creditors/{creditorId}:
    parameters:
      - $ref: '#/components/parameters/ApplicationIdPath'
      - $ref: '#/components/parameters/CreditorIdPath'
    get:
      tags: [Creditors]
      summary: Get a creditor
      responses:
        '200':
          description: The creditor
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Creditors]
      summary: Edit a creditor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreditorRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreditorResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Creditors]
      summary: Delete a creditor
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Assets
  /api/v1/applications/{applicationId}/assets:
    parameters:
      - $ref: '#/components/parameters/ApplicationIdPath'
    get:
      tags: [Assets]
      summary: List assets for a case
      responses:
        '200':
          description: Assets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AssetResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Assets]
      summary: Add an asset (duplicates are flagged, not rejected)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/applications/{applicationId}/assets/{assetId}:
    parameters:
      - $ref: '#/components/parameters/ApplicationIdPath'
      - $ref: '#/components/parameters/AssetIdPath'
    get:
      tags: [Assets]
      summary: Get an asset
      responses:
        '200':
          description: The asset
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Assets]
      summary: Edit an asset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Assets]
      summary: Delete an asset
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ApplicationIdPath:
      name: applicationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CreditorIdPath:
      name: creditorId
      in: path
      required: true
      schema: { type: string, format: uuid }
    AssetIdPath:
      name: assetId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Authenticated but not permitted (role or ownership)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
...
    ApiError:
      type: object
      description: >-
        Uniform error payload returned by every endpoint on failure. `fieldErrors`
        is omitted when empty. `code` is a stable client-facing value; framework
        and database details are never exposed.
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 400 }
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, example: Validation failed }
        path: { type: string, example: /api/v1/auth/register }
        correlationId:
          type: string
          description: Echoes the request's X-Correlation-Id.
        fieldErrors:
          type: array
          items: { $ref: '#/components/schemas/ApiFieldError' }
      required: [timestamp, status, code, message, path, correlationId]
    ApiFieldError:
      type: object
      properties:
        field: { type: string, example: email }
        message: { type: string, example: must be a well-formed email address }
      required: [field, message]
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - RESOURCE_NOT_FOUND
        - ACCESS_DENIED
        - INVALID_STATE
        - QUESTIONNAIRE_VALIDATION_ERROR
        - RULES_EVALUATION_ERROR
        - DOCUMENT_GENERATION_ERROR
        - AUTHENTICATION_FAILED
        - INTERNAL_ERROR
...
    CreditorType:
      type: string
      enum: [BANK, MICROFINANCE, INDIVIDUAL, TAX_AUTHORITY, UTILITY, OTHER]
    AssetType:
      type: string
      enum: [REAL_ESTATE, VEHICLE, BANK_ACCOUNT, SECURITIES, CASH, RECEIVABLE, MOVABLE_PROPERTY, OTHER]
...
    CreditorRequest:
      type: object
      properties:
        name: { type: string, maxLength: 300 }
        type: { $ref: '#/components/schemas/CreditorType' }
        inn: { type: string, maxLength: 12, nullable: true }
        claimBasis: { type: string, maxLength: 500, nullable: true }
        currency:
          type: string
          maxLength: 3
          nullable: true
          description: ISO-4217 code; defaults to RUB when omitted.
        totalAmount: { type: number, format: double, minimum: 0 }
        principalAmount: { type: number, format: double, minimum: 0, nullable: true }
        interestAmount: { type: number, format: double, minimum: 0, nullable: true }
        penaltyAmount: { type: number, format: double, minimum: 0, nullable: true }
        overdue: { type: boolean, default: false }
        secured: { type: boolean, default: false }
      required: [name, type, totalAmount]
    CreditorResponse:
      type: object
      properties:
        creditorId: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        name: { type: string }
        type: { $ref: '#/components/schemas/CreditorType' }
        inn: { type: string, nullable: true }
        claimBasis: { type: string, nullable: true }
        currency: { type: string }
        totalAmount: { type: number, format: double }
        principalAmount: { type: number, format: double, nullable: true }
        interestAmount: { type: number, format: double, nullable: true }
        penaltyAmount: { type: number, format: double, nullable: true }
        overdue: { type: boolean }
        secured: { type: boolean }
        duplicateWarning:
          type: boolean
          description: True when another creditor in the case looks like a duplicate (non-blocking).
      required: [creditorId, applicationId, name, type, currency, totalAmount, overdue, secured, duplicateWarning]

    # ---- Estate: assets ----
    AssetRequest:
      type: object
      properties:
        type: { $ref: '#/components/schemas/AssetType' }
        description: { type: string, maxLength: 500 }
        currency:
          type: string
          maxLength: 3
          nullable: true
          description: ISO-4217 code; defaults to RUB when omitted.
        estimatedValue: { type: number, format: double, minimum: 0 }
        ownershipShare: { type: string, maxLength: 16, nullable: true }
        registrationNumber: { type: string, maxLength: 100, nullable: true }
        acquiredOn: { type: string, format: date, nullable: true }
        pledged: { type: boolean, default: false }
        pledgeCreditorId: { type: string, format: uuid, nullable: true }
      required: [type, description, estimatedValue]
    AssetResponse:
      type: object
      properties:
        assetId: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/AssetType' }
        description: { type: string }
        currency: { type: string }
        estimatedValue: { type: number, format: double }
        ownershipShare: { type: string, nullable: true }
        registrationNumber: { type: string, nullable: true }
        acquiredOn: { type: string, format: date, nullable: true }
        pledged: { type: boolean }
        pledgeCreditorId: { type: string, format: uuid, nullable: true }
        duplicateWarning:
          type: boolean
          description: True when another asset in the case looks like a duplicate (non-blocking).
      required: [assetId, applicationId, type, description, currency, estimatedValue, pledged, duplicateWarning]