# src/main/java/ru/adiaphora/platform/estate/domain/AssetType.java
package ru.adiaphora.platform.estate.domain;

/**
 * Kind of asset in the debtor's estate. Placeholder taxonomy pending lawyer review.
 */
public enum AssetType {
    REAL_ESTATE,
    VEHICLE,
    BANK_ACCOUNT,
    SECURITIES,
    CASH,
    RECEIVABLE,
    MOVABLE_PROPERTY,
    OTHER
}